+++ _spy/vitollino/__init__.py
@@ -10,5 +10,4 @@ Changelog
         version.
 
 """
-__version__ = "20.08.b0"