+++ kwarwp/__log__.py
@@ -107,3 +107,245 @@ AttributeError: 'Vitollino' object has no attribute 'c'
     indio = self.v.e(self.INDIO, cena=cena)
 AttributeError: 'Jogo' object has no attribute 'e'
 '''},
+{'date': 'Wed Aug 19 2020 21:05:44.449 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 396
+    main(Jogo, STYLE)
+  module <module> line 390
+    Kwarwp(vitollino, medidas=medidas)
+  module <module> line 236
+    self.cena = self.cria(mapa=self.mapa) if vitollino else None
+  module <module> line 270
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 338
+    coisa = Piche(imagem, x=0, y=0, cena=cena, taba=self)
+  module kwarwp.kwarwpart line 152
+    self.posicao = (x//lado,y//lado-1)
+TypeError: unsupported operand type(s) for //: 'int' and 'NoneType'
+'''},
+{'date': 'Wed Aug 19 2020 21:09:00.567 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 396
+    main(Jogo, STYLE)
+  module <module> line 390
+    Kwarwp(vitollino, medidas=medidas)
+  module <module> line 236
+    self.cena = self.cria(mapa=self.mapa) if vitollino else None
+  module <module> line 270
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 338
+    coisa = Piche(imagem, x=0, y=0, cena=cena, taba=self)
+  module kwarwp.kwarwpart line 152
+    self.posicao = (x//lado,y//lado-1)
+TypeError: unsupported operand type(s) for //: 'int' and 'NoneType'
+'''},
+{'date': 'Wed Aug 19 2020 21:12:37.628 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 396
+    main(Jogo, STYLE)
+  module <module> line 390
+    Kwarwp(vitollino, medidas=medidas)
+  module <module> line 236
+    self.cena = self.cria(mapa=self.mapa) if vitollino else None
+  module <module> line 270
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 338
+    coisa = Piche(imagem, x=0, y=0, cena=cena, taba=self)
+  module kwarwp.kwarwpart line 153
+    self.vazio = Kwarwp.VITOLLINO.a(imagem, w=lado, h=lado, x=0, y=0, cena=cena)
+AttributeError: 'NoneType' object has no attribute 'a'
+'''},
+{'date': 'Wed Aug 19 2020 21:18:46.81 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 396
+    main(Jogo, STYLE)
+  module <module> line 390
+    Kwarwp(vitollino, medidas=medidas)
+  module <module> line 236
+    self.cena = self.cria(mapa=self.mapa) if vitollino else None
+  module <module> line 270
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 338
+    coisa = Piche(imagem, x=0, y=0, cena=cena, taba=self)
+  module kwarwp.kwarwpart line 155
+    self.vazio = Kwarwp.VITOLLINO.a(imagem, w=lado, h=lado, x=0, y=0, cena=cena)
+AttributeError: 'Jogo' object has no attribute 'a'
+'''},
+{'date': 'Wed Aug 19 2020 21:20:56.285 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 396
+    main(Jogo, STYLE)
+  module <module> line 390
+    Kwarwp(vitollino, medidas=medidas)
+  module <module> line 236
+    self.cena = self.cria(mapa=self.mapa) if vitollino else None
+  module <module> line 270
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 339
+    vaga = Vazio("", x=x, y=y, cena=cena, ocupante=coisa)
+  module kwarwp.kwarwpart line 52
+    self.posicao = (x//lado,y//lado-1)
+TypeError: unsupported operand type(s) for //: 'int' and 'NoneType'
+'''},
+{'date': 'Fri Aug 21 2020 08:37:14.916 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 396
+    main(Jogo, STYLE)
+  module <module> line 390
+    Kwarwp(vitollino, medidas=medidas).cria()
+  module <module> line 270
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 339
+    vaga = Vazio("", x=x, y=y, cena=cena, ocupante=coisa)
+  module kwarwp.kwarwpart line 54
+    self.posicao = (x//lado,y//lado-1)
+TypeError: unsupported operand type(s) for //: 'int' and 'NoneType'
+'''},
+{'date': 'Fri Aug 21 2020 09:27:04.656 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 397
+    main(Jogo, STYLE)
+  module <module> line 391
+    Kwarwp(vitollino, medidas=medidas).cria()
+  module <module> line 234
+    kmain(self)
+NameError: name 'kmain' is not defined
+'''},
+{'date': 'Fri Aug 21 2020 09:27:19.822 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 397
+    main(Jogo, STYLE)
+  module <module> line 391
+    Kwarwp(vitollino, medidas=medidas).cria()
+  module <module> line 271
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 374
+    self.o_indio = Indio(imagem, x=1, y=0, cena=cena, taba=self)
+  module <module> line 74
+    self.posicao = (x//lado,y//lado)
+TypeError: unsupported operand type(s) for //: 'int' and 'NoneType'
+'''},
+{'date': 'Fri Aug 21 2020 09:29:08.198 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 397
+    main(Jogo, STYLE)
+  module <module> line 391
+    Kwarwp(vitollino, medidas=medidas).cria()
+  module <module> line 271
+    self.taba = {(i, j): fabrica[imagem].objeto(fabrica[imagem].imagem, x=i*lado, y=j*lado+lado, cena=cena)
+  module <module> line 374
+    self.o_indio = Indio(imagem, x=1, y=0, cena=cena, taba=self)
+  module <module> line 75
+    self.indio = Kwarwp.VITOLLINO.a(imagem, w=lado, h=lado, x=x, y=y, cena=cena)
+AttributeError: 'NoneType' object has no attribute 'a'
+'''},
+{'date': 'Sat Aug 22 2020 09:25:30.675 GMt-0300 (Horário Padrão de Brasília) -X- SuPyGirls -X-',
+'error': '''Traceback (most recent call last):
+  module _core.main line 177
+    dialog.action(lambda *_: self.start()
+  module _core.supygirls_factory line 135
+    self.act(self, lambda *_: self.hide() or extra()) if self.act else None
+  module _core.supygirls_factory line 306
+    return self._first_response(lambda: self._executa_acao(), self.extra, self.error)
+  module _core.supygirls_factory line 278
+    traceback.print_exc(file=sys.stderr)
+  module _core.supygirls_factory line 295
+    exec(self.code, glob)  # dict(__name__="__main__"))
+  module <module> line 400
+    main(Jogo, STYLE)
+  module <module> line 394
+    Kwarwp(vitollino, medidas=medidas).cria()
+  module <module> line 249
+    fabrica = {
+AttributeError: 'Kwarwp' object has no attribute 'indi0'
+'''},