+++ CHANGELOG.rst
@@ -12,8 +12,46 @@ The format is based on `Keep a Changelog`_.
 -------------
 - to be defined
 
+`20.08.b0`_
+------------
+
+Added
++++++
+- Actions pick and drop.
+- Class Tora.
+
+`20.08.a3`_
+-----------
+
+Added
++++++
+- Actions turn left and right.
+- Classes Piche e Oca.
+
+`20.08.a2`_
+-----------
+
+Added
++++++
+- class Vazio, Coisa, entra action.
+
+`20.08.a1`_
+-----------
+
+Added
++++++
+- Indian class add moves and execs
+
+`20.08.a0`_
+-----------
+
+Added
++++++
+- Code and libs for deploy.
+- Indian class at kwarapp
+
 `20.07`_
---------
+---------
 
 Added
 +++++
@@ -38,6 +76,12 @@ LABASE_ - NCE_ - UFRJ_
 .. _NCE: http://nce.ufrj.br
 .. _UFRJ: http://www.ufrj.br
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0/
+.. _20.07: https://github.com/kwarwp/kwarwp/releases/tag/20.07
+.. _20.08.a0: https://github.com/kwarwp/kwarwp/releases/tag/20.08.a0
+.. _20.08.a1: https://github.com/kwarwp/kwarwp/tree/release_20_08_a1
+.. _20.08.a2: https://github.com/kwarwp/kwarwp/tree/release_20_08_a2
+.. _20.08.a3: https://github.com/kwarwp/kwarwp/tree/release_20_08_a3
+.. _20.08.b0: https://github.com/kwarwp/kwarwp/tree/release_20_08_b0
 
 .. |LABASE| image:: https://cetoli.gitlab.io/spyms/image/labase-logo-8.png
    :target: http://labase.activufrj.nce.ufrj.br